fixed Delete request issue

+++ src/components/Manager.jsx
@@ -13,7 +13,7 @@ const Manager = () => {
   const getPasswords = async () => {
     let req = await fetch("http://localhost:3000");
     let passwords = await req.json();
-    console.log(passwords);
+    // console.log(passwords);
     setpasswordArray(passwords);
   };
 
@@ -44,15 +44,16 @@ const Manager = () => {
       form.uname.length > 3 &&
       form.password.length > 3
     ) {
-      await fetch("http://localhost:3000", {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-        },
-
-        body: JSON.stringify({ id: form.id }),
-      });
+      if (form.id) {
+        await fetch(`http://localhost:3000/${form.id}`, {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+          },
+        });
+      }
       setpasswordArray([...passwordArray, { ...form, id: uuidv4() }]);
+
       let res = await fetch("http://localhost:3000", {
         method: "POST",
         headers: {
